requirements.txt
aniso8601==9.0.1
asgiref==3.2.10
Babel==2.10.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
ClickUz==0.3
colorama==0.4.4
confusable-homoglyphs==3.2.0
cryptography==36.0.2
decorator==4.4.2
defusedxml==0.7.1
Django==3.0.9
django-admin-thumbnails==0.2.5
django-allauth==0.50.0
django-ckeditor==6.0.0
django-classy-tags==2.0.0
django-cms==3.8.0
django-cors-headers==3.7.0
django-crispy-forms==1.14.0
django-currencies==0.10.1
django-decorator-include==3.0
django-discover-runner==1.0
django-filer==2.0.2
django-filter==21.1
django-formtools==2.2
django-hosts==5.1
django-js-asset==1.2.2
django-money==2.1.1
django-mptt==0.11.0
django-polymorphic==3.0.0
django-push==1.1
django-ratelimit==3.0.1
django-recaptcha==3.0.0
django-registration==2.4.1
django-resized==0.3.11
django-restframework==0.0.1
django-role-permissions==3.1.1
django-sekizai==2.0.0
django-treebeard==4.3.1
djangocms-admin-style==2.0.1
djangocms-attributes-field==2.0.0
djangocms-bootstrap4==2.0.0
djangocms-icon==2.0.0
djangocms-link==3.0.0
djangocms-picture==3.0.0
djangocms-text-ckeditor==4.0.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
docutils==0.18.1
easy-thumbnails==2.7
feedparser==6.0.8
graphene==3.0
graphene-django==3.0.0b7
graphql-core==3.1.7
graphql-relay==3.1.5
html5lib==1.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.4
jsonfield==2.1.1
moviepy==1.0.3
mysqlclient==2.0.3
numpy==1.21.2
oauthlib==3.2.0
PaycomUz==2.2
Pillow==7.1.2
proglog==0.1.9
promise==2.3
psycopg2-binary==2.9.3
py-moneyed==1.2
pycparser==2.20
PyJWT==1.7.1
pyTelegramBotAPI==3.7.5
python-dotenv==0.18.0
python-ffmpeg-video-streaming==0.1.14
python-gettext==4.0
python3-openid==3.2.0
pytz==2020.1
requests==2.25.1
requests-oauthlib==1.3.1
sgmllib3k==1.0.0
six==1.15.0
sorl-thumbnail==12.8.0
soupsieve==2.0.1
sqlparse==0.3.1
stripe==2.74.0
text-unidecode==1.3
tqdm==4.62.1
transliterate==1.10.2
tzdata==2022.1
Unidecode==1.1.1
urllib3==1.26.2
webencodings==0.5.1
